% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empires_size_3000_600}
\alias{empires_size_3000_600}
\title{DATASET_TITLE}
\format{
A data frame with 118 rows and 9 variables:
\describe{
\item{\code{Year}}{integer COLUMN_DESCRIPTION}
\item{\code{Area}}{double COLUMN_DESCRIPTION}
\item{\code{Error}}{double COLUMN_DESCRIPTION}
\item{\code{Notes}}{character COLUMN_DESCRIPTION}
\item{\code{Empire}}{character COLUMN_DESCRIPTION}
\item{\code{Region}}{character COLUMN_DESCRIPTION}
\item{\code{Min}}{double COLUMN_DESCRIPTION}
\item{\code{Max}}{double COLUMN_DESCRIPTION}
\item{\code{Average}}{double COLUMN_DESCRIPTION}
}
}
\usage{
empires_size_3000_600
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
