% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo_stations}
\alias{meteo_stations}
\title{DATASET_TITLE}
\format{
A data frame with 23 rows and 5 variables:
\describe{
\item{\code{station}}{character COLUMN_DESCRIPTION}
\item{\code{location}}{character COLUMN_DESCRIPTION}
\item{\code{alt_ft}}{double COLUMN_DESCRIPTION}
\item{\code{lat}}{double COLUMN_DESCRIPTION}
\item{\code{long}}{double COLUMN_DESCRIPTION}
}
}
\usage{
meteo_stations
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
