% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_mesopotamia_sites_NJS}
\alias{n_mesopotamia_sites_NJS}
\title{DATASET_TITLE}
\format{
A data frame with 1041 rows and 8 variables:
\describe{
\item{\code{Id}}{character COLUMN_DESCRIPTION}
\item{\code{SizeHa}}{double COLUMN_DESCRIPTION}
\item{\code{StartDate}}{double COLUMN_DESCRIPTION}
\item{\code{EndDate}}{double COLUMN_DESCRIPTION}
\item{\code{Longitude}}{double COLUMN_DESCRIPTION}
\item{\code{Latitude}}{double COLUMN_DESCRIPTION}
\item{\code{Period}}{character COLUMN_DESCRIPTION}
\item{\code{geom}}{list COLUMN_DESCRIPTION}
}
}
\usage{
n_mesopotamia_sites_NJS
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
