% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tellspotting_sites}
\alias{tellspotting_sites}
\title{DATASET_TITLE}
\format{
A data frame with 532 rows and 2 variables:
\describe{
\item{\code{elevation}}{double COLUMN_DESCRIPTION}
\item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\usage{
tellspotting_sites
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
